changed app folder name to app. automated db creation.
