+++ app/templates/login.html
@@ -1,0 +1,120 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <title>Log in Form</title>
+    <meta charset="UTF-8">
+    <style>
+        body {
+            justify-content: center;
+            align-items: center;
+            background-image: url("../static/image.jpg");
+            background-size: cover;
+            color: #f5f6f7;
+            font-family: 'Tahoma';
+            font-size: 16px;
+            font-weight: bold;
+
+        }
+
+        h1 {
+            margin: 1em auto;
+            text-align: center;
+            text-decoration-line: underline;
+            font-weight: bolder;
+
+        }
+
+        p {
+            margin: 1em auto;
+            text-align: center;
+            text-decoration-line: underline;
+            font-weight: bolder;
+
+
+        }
+
+        form {
+            margin: 0 auto;
+            max-width: 500px;
+            min-width: 300px;
+            width: 60vw;
+            padding-bottom: 2em;
+
+        }
+
+        fieldset {
+            border-color: #ff0cff;
+            border-style: dashed;
+            padding-top: 2rem;
+            padding-bottom: 2rem;
+            padding-left: 0;
+            padding-right: 0;
+
+        }
+
+        label {
+
+            display: block;
+            width: fit-content;
+            margin-left: 5px;
+        }
+
+        input {
+            width: 100%;
+            margin-top: 10px;
+            margin-bottom: 0;
+            margin-right: 0;
+            margin-left: 5px;
+            background-color: #D9D9D9;
+            border: 1px solid;
+            min-height: 0.7em;
+        }
+
+        button {
+            display: inline-block;
+            width: 20%;
+            font-size: medium;
+
+        }
+        .content_div{
+            color: rgb(255, 255, 255);
+            background-color: rgba(34, 18, 39, 0.5);
+            background-size:contain;
+            width: 525px;
+            margin:auto;
+            border: 5px;
+            border-width: 4px;
+            border-color: coral;
+            margin-top: 150px;
+        }
+    </style>
+</head>
+
+<body>
+    <div class="content_div">
+        <h1> Log In</h1>
+        <p> Please fill out your details to log in</p>
+
+        <form action="{{url_for('login')}}" method="post">
+            <fieldset>
+                <label for="email">Enter Your Email: <input type="email" id="email" name="email" required>
+                </label>
+
+                <label for="user_password">Enter Your Password: <input type="password" id="user_password"
+                        name="user_password" required minlength="6">
+                </label>
+                <p>{{error_msg}}</p>
+            </fieldset><br>
+            <button type="submit" onclick="history.back()">Back</button>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+            &nbsp;&nbsp; &nbsp;&nbsp; &nbsp; &nbsp; &nbsp;&nbsp; &nbsp;&nbsp; &nbsp;&nbsp; &nbsp;&nbsp;
+            &nbsp;&nbsp;&nbsp;
+            &nbsp;&nbsp; &nbsp; &nbsp;&nbsp; &nbsp;&nbsp;&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
+            &nbsp;<button type="submit">Log In</button>
+
+        </form>
+    </div>
+
+</body>
+
+</html>
