+++ app/templates/register.html
@@ -1,0 +1,179 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <title>Registration Form</title>
+    <meta charset="UTF-8">
+    <style>
+        body {
+            justify-content: center;
+            align-items: center;
+            background-image: url("../static/image.jpg"); 
+            background-size: cover;
+            color: #f5f6f7;
+            font-family: 'Tahoma';
+            font-size: 16px;
+            font-weight: bold;
+
+        }
+
+        h1 {
+            margin: 1em auto;
+            text-align: center;
+            text-decoration-line: underline;
+            font-weight: bolder;
+
+        }
+
+        p {
+            margin: 1em auto;
+            text-align: center;
+            font-weight: bolder;
+
+        }
+
+        form {
+            margin: 0 auto;
+            max-width: 500px;
+            min-width: 300px;
+            width: 60vw;
+            padding-bottom: 2em;
+
+        }
+
+        fieldset {
+            border-color: #ff0cff;
+            border-width: small;
+            border-style: dashed;
+            border-spacing:20px;
+            padding-top: 2rem;
+            padding-bottom: 2rem;
+            padding-left: 0;
+            padding-right: 0;
+
+        }
+
+        label {
+            display: block;
+            width: fit-content;
+            margin-left: 5px;
+        }
+
+        input,
+        textarea,
+        select {
+            width: 100%;
+            margin-top: 10px;
+            margin-bottom: 0;
+            margin-right: 0;
+            margin-left: 0;
+            max-width: fit-content;
+        }
+
+        .inline {
+            width: unset;
+            margin-right: 0.5em;
+            margin-left: 0;
+            margin-top: 0;
+            margin-bottom: 0;
+            vertical-align: middle;
+        }
+
+        input {
+            background-color: #D9D9D9;
+            border: 1px solid;
+            min-height: 0.7em;
+            margin-left: 5px;
+            max-width: fit-content;
+
+        }
+
+        textarea {
+            background-color: rgba(217, 217, 217);
+            
+            border: 1px solid;
+            min-height: 0.7em;
+            max-width: fit-content;
+            margin-left: auto;
+        }
+
+        button {
+            display: inline-block;
+            width: 20%;
+            font-size: medium;
+            position: relative;
+        }
+        .content_div{
+            color: rgb(255, 255, 255);
+            background-color: rgba(34, 18, 39, 0.5);
+            background-size:contain;
+            width: 525px;
+            margin:auto;
+            border: 5px;
+            border-width: 4px;
+            border-color: coral;
+        }
+    </style>
+</head>
+
+<body>
+    <div class="content_div">
+        <h1> Registration Form</h1>
+        <p> Please fill out the registration form with required information</p>
+
+        <form action="{{url_for('register')}}" method="post">
+            <fieldset>
+                <label for="first-name">Enter Your First Name: <input type="text" id="first-name" name="first-name"
+                        required>
+                </label>
+
+                <label for="last-name">Enter Your Last Name: <input type="text" id="last-name" name="last-name"
+                        required>
+                </label>
+
+                <label for="email">Enter Your Email: <input type="email" id="email" name="email" required>
+                </label>
+
+                <label for="new-password">Create a New Password: <input type="password" id="new-password"
+                        name="new-password" required minlength="6">
+                </label>
+            </fieldset>
+
+            <fieldset>
+                <label for="coach"><input type="radio" name="user-type" id="coach" value="optionCoach" class="inline">
+                    Coach</label>
+                <label for="coachee"><input type="radio" name="user-type" id="coachee" value="optionCoachee"
+                        class="inline"> Coachee</label>
+                <label for="tcsandcs"><input type="checkbox" name="tcs" required id="tcsandcs" class="inline"> <a> I
+                        accept
+                        the <a href="http://localhost:5000/terms-and-conditions">terms and conditions</a></label>
+            </fieldset>
+
+            <fieldset>
+                <label for="reference"> How did you hear about us?
+                    <select name="" id="reference">
+                        <option value="">(select one)</option>
+                        <option value="1">TV advertisement</option>
+                        <option value="2">Social Media</option>
+                        <option value="3">Word of mouth</option>
+                        <option value="4">Other</option>
+                    </select>
+                </label>
+
+                <label for="moreInfo">Optional info you want to tell us about yourself:
+                    <textarea id="moreInfo" rows="3" cols="30" placeholder="I like coding..."></textarea>
+                </label>
+            </fieldset><br>
+            <button type="submit" onclick="history.back()">Back</button>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+            &nbsp;&nbsp; &nbsp;&nbsp; &nbsp; &nbsp; &nbsp;&nbsp; &nbsp;&nbsp; &nbsp;&nbsp; &nbsp;&nbsp;
+            &nbsp;&nbsp;&nbsp; &nbsp;&nbsp; &nbsp; &nbsp;&nbsp; &nbsp;&nbsp;&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
+            &nbsp; &nbsp; &nbsp; &nbsp;<button type="submit">Register</button>
+
+
+
+        </form>
+    </div>
+
+</body>
+
+</html>
