--- register_and_login/readme.md
@@ -1,4 +1,0 @@
-updated query file to reflect changes in rating system database <br>
-updated get_credentials_tuple() to now accept email address instead of username <br>
-test_app.py now runs checks based on email <br>
-login.html and register.html updated accordingly <br>
